Fix run-tests.sh so error codes are propagated
`set -e` is required so that the error code from the parallel command is
propagated to the parent shell.

`set -o pipefail` is required in the parallel command so that the error
code from `./all-tests.sh` is propagated to the parent shell. Otherwise
this error code is masked by `| tee`.